Structure Refactoring: Phase 01
