+++ lib/steps/step_03_remove_duplicates/step_03_remove_duplicates.dart
@@ -1,4 +1,4 @@
-import '../models/pipeline_step_model.dart';
+import '../../domain/models/pipeline_step_model.dart';
 
 /// Step 3: Remove duplicate media files
 ///
