+++ lib/steps/step_08_update_creation_time/step_08_update_creation_time.dart
@@ -6,8 +6,8 @@ import 'package:console_bars/console_bars.dart';
 import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
 
-import '../models/pipeline_step_model.dart';
-import '../services/core/logging_service.dart';
+import '../../domain/models/pipeline_step_model.dart';
+import '../../domain/services/core/logging_service.dart';
 
 /// Step 8: Update creation times (Windows only)
 ///
