+++ test/e2e/e2e_github_issues_test.dart
@@ -23,7 +23,7 @@ library;
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:gpth/domain/main_pipeline.dart';
+import 'package:gpth/steps/main_pipeline.dart';
 import 'package:gpth/domain/models/processing_config_model.dart';
 import 'package:gpth/domain/services/core/service_container.dart';
 import 'package:gpth/domain/services/user_interaction/path_resolver_service.dart';
