+++ test/e2e/e2e_missing_coverage_test.dart
@@ -14,7 +14,7 @@ library;
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:gpth/domain/main_pipeline.dart';
+import 'package:gpth/steps/main_pipeline.dart';
 import 'package:gpth/domain/models/processing_config_model.dart';
 import 'package:gpth/domain/services/core/service_container.dart';
 import 'package:gpth/domain/services/user_interaction/path_resolver_service.dart';
